Migrate app from Javascript to Typescript #21
