+++ src/components/ViewProfile/container/ProfileSummary.tsx
@@ -1,14 +1,85 @@
-import React from 'react';
+import React, { FC, useEffect, useState } from 'react';
+import { useSelector } from 'react-redux';
+import { History, LocationState } from 'history';
 
-const ProfileSummary = ({
-    avatar,
-    name,
-    username,
-    location,
-    followers,
-    following,
-    repositories,
-}) => {
+interface Props{
+    history: History<LocationState>;
+}
+
+interface stateProps{
+    profile: {
+        profile: {
+            avatar_url: string;
+            name: string;
+            email: string;
+            location: string;
+            public_repos: number;
+            followers: number;
+            following: number;
+            login: string
+        };
+        loading: boolean;
+    };
+    repos: {
+        repos: [{
+            language: string;
+            name: string;
+            stargazers_count: number;
+        }];
+        loading: boolean;
+    };
+}
+const ProfileSummary: FC<Props> = ({ history }) => {
+    const [user, setUser] = useState({
+        avatar: '',
+        name: '',
+        username: '',
+        location: '',
+        repositories: 0,
+        followers: 0,
+        following: 0,
+    });
+
+    const { profile, repos } = useSelector(
+        (state: stateProps) => state,
+    );
+
+    useEffect(() => {
+        if (repos.repos === null && !repos.loading) {
+            history.push('/');
+        } else {
+            const {
+                avatar_url,
+                name,
+                login,
+                location,
+                public_repos,
+                followers,
+                following,
+            } = profile.profile;
+
+            setUser({
+                avatar: avatar_url,
+                name,
+                username: login,
+                location,
+                repositories: public_repos,
+                followers,
+                following,
+            });
+        }
+    }, [history, repos.repos, repos.loading, profile.profile]);
+
+    const {
+        avatar,
+        name,
+        username,
+        location,
+        repositories,
+        followers,
+        following,
+    } = user;
+
     return (
         <div className="profile__summary">
             <div className="container">
@@ -29,7 +100,8 @@ const ProfileSummary = ({
                         <div className="location">
                             <span>
                                 <i className="fas fa-map-marker" />
-                            </span> &nbsp; &nbsp;
+                            </span>{' '}
+                            &nbsp; &nbsp;
                             {location}
                         </div>
                         <div className="flex__container">
