--- src/config/config.js
@@ -1,4 +1,0 @@
-export const baseUrl = 'https://api.github.com/users';
-export const clientId = 'Iv1.3257a9819dfbb01e';
-
-export const clientSecret = 'f5f4e39e3190e9c1f9d04c829306d3dca9f7679f';
