+++ src/config/config.tsx
@@ -1,0 +1,4 @@
+export const baseUrl: string = 'https://api.github.com/users';
+export const clientId: string = 'Iv1.3257a9819dfbb01e';
+
+export const clientSecret: string = 'f5f4e39e3190e9c1f9d04c829306d3dca9f7679f';
