+++ src/helpers/langColors.tsx
@@ -1,4 +1,4 @@
-const langColors = {
+const langColors: {[k: string]: string} = {
     Mercury: '#ff2b2b',
     TypeScript: '#2b7489',
     PureBasic: '#5a6986',
