+++ src/redux/reducers/baseAction.tsx
@@ -1,0 +1,4 @@
+export interface BaseAction {
+    type: string;
+    payload?: object;
+}
