+++ src/redux/reducers/mainReducer.tsx
