+++ src/redux/reducers/profileReducer.tsx
@@ -1,11 +1,12 @@
 import { FETCH_PROFILE, PROFILE_LOADING } from '../actions/types';
+import {BaseAction} from './baseAction';
 
-const initialState = {
+export const initialState = {
     profile: null,
     loading: false,
 };
 
-const profileReducer = (state = initialState, action) => {
+const profileReducer = (state = initialState, action: BaseAction) => {
     switch (action.type) {
         case PROFILE_LOADING:
             return {
